pyproject.toml
[tool.poetry]
name = "fastpoet"
version = "0.1.0"
description = ""
authors = ["Иван Алексеев <seevvia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.36"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
flake8 = "^4.0.1"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
